+++ src/Entities/GpApi/GpApiAccount.php
@@ -4,12 +4,14 @@ namespace GlobalPayments\Api\Entities\GpApi;
 
 class GpApiAccount
 {
-    public $id;
-    public $name;
+    public ?string $id;
+    public ?string $name;
+    public ?array $permissions;
 
-    public function __construct($id, $name)
+    public function __construct(?string $id, ?string $name, ?array $permissions = null)
     {
         $this->id = $id;
         $this->name = $name;
+        $this->permissions = $permissions;
     }
 }